move helper functions into their own file
