+++ client/src/pages/EventPage.tsx
@@ -11,6 +11,7 @@ import { createEvent } from "../api/apiService";
 import TextArea from "../components/TextArea";
 import Toast, { ToastProps } from "../components/Toast";
 import { useSearchParams } from "react-router";
+import { formatDate, convertDateToInputFormat, formatTime } from "../utils/formatUtils";
 
 const EventPage = () => {
     const [isCreateModalOpen, setIsCreateModalOpen] = useState(false);
@@ -598,44 +599,6 @@ const EventTable = ({ openCreateEventModal, setToast, toastVisible, setToastVisi
             </div>
         </div>
     );
-};
-
-const formatDate = (date: Date) => {
-    // TODO should this be a utility function? (moved to utils dir)
-    const formatter = new Intl.DateTimeFormat("en-US", {
-        weekday: "short",
-        year: "numeric",
-        month: "short",
-        day: "2-digit",
-    });
-
-    return formatter.format(date);
-}
-
-const formatTime = (date: Date) => {
-    const formatter = new Intl.DateTimeFormat("en-US", {
-        hour: "2-digit",
-        minute: "2-digit",
-        hour12: true
-    });
-
-    return formatter.format(date);
-};
-
-const convertDateToInputFormat = (date: Date | string) => {
-    // Ensure the date is of type Date;
-    if (typeof date === "string") {
-        date = new Date(date);
-    }
-
-    // Account for timezone offset
-    const timezoneOffset = date.getTimezoneOffset() * 60000; // Convert minutes to milliseconds
-    const adjustedDate = new Date(date.getTime() - timezoneOffset);
-
-    // Format the date to match the input type="datetime-local" format
-    const formattedDate = adjustedDate.toISOString().slice(0, 16);
-
-    return formattedDate;
 };
 
 export default EventPage;
